Updated FileOutput plugin to address concurrency and performance concerns.
I see no reason to make the entire onDataReady method synchronized, when the
real issue is that we don't want files to be created more than once if two
packets of the same datatype come in at the same time. So I wrapped the part
where the files are created in a synchronized block. This should improve
performance, as it allows multiple threads to be logging concurrently,
instead of having the entire method require exclusive access.